shrunk info.c table a whole bunch again, moved some larger sparsely populated (and accessed)  fields to getters with a small switch block. there are a few more other candidates in there.. along the way changed the way we handled some fields like speed and radius and height to be more potentitally 16 bit friendly later.

+++ sounds.h
@@ -73,13 +73,13 @@ typedef struct
     int8_t*	name;
 
     // lump number of music
-    int32_t		lumpnum;
+    int16_t		lumpnum;
     
     // music data
     void*	data;
 
     // music handle once registered
-	int32_t handle;
+	int16_t handle;
     
 } musicinfo_t;
